Move run command into own file
