+++ cli.js
@@ -465,75 +465,7 @@ yargs
   .command(require('./commands/mint'))
   .command(require('./commands/ls'))
   .command(require('./commends/count'))
-  .command('run [id]', 'run a package in browser', yargs => {
-    yargs
-      .option('path', {
-        alias: 'p',
-        type: 'string',
-        description: 'Use local xrpackage path for runtime',
-      });
-  }, async argv => {
-    handled = true;
-
-    const app = express();
-    app.use((req, res, next) => {
-      res.set('Access-Control-Allow-Origin', '*');
-      res.set('Access-Control-Allow-Methods', '*');
-      res.set('Access-Control-Allow-Headers', '*');
-      next();
-    });
-
-    const _getRunUrl = o => {
-      let url;
-      let servePath;
-      if (o.path) {
-        url = `http://localhost:${port}/run.html`;
-        servePath = o.path;
-      } else {
-        url = 'https://xrpackage.org/run.html';
-        servePath = null;
-      }
-      if (o.id) {
-        url += `?i=${o.id}`;
-      } else if (o.url) {
-        url += `?u=${o.url}`;
-      }
-      return {
-        url,
-        servePath,
-      };
-    };
-
-    let runSpec;
-    if (!isNaN(parseInt(argv.id, 10))) {
-      runSpec = _getRunUrl({
-        path: argv.path,
-        id: argv.id,
-      });
-    } else {
-      runSpec = _getRunUrl({
-        path: argv.path,
-        url: `http://localhost:${port}/a.wbn`,
-      });
-      app.get('/a.wbn', (req, res) => {
-        const rs = fs.createReadStream(argv.id);
-        rs.pipe(res);
-        rs.once('error', err => {
-          res.statusCode = 500;
-          res.end(err.stack);
-        });
-      });
-    }
-    const {url, servePath} = runSpec;
-    if (servePath) {
-      app.use(express.static(servePath));
-    }
-
-    const server = http.createServer(app);
-    server.listen(port, () => {
-      open(url);
-    });
-  })
+  .command(require('./commands/run'))
   .command('inspect [id]', 'inspect a package in browser', yargs => {
     yargs
       .option('path', {
